Fix AMPlainFAPIConsentHandler handling of authorization_code grant flow
The handler implements authorization_code grant flow in the same way that the AM UI works, except that it is being used in a headless fashion.

The consent decision call is made to the /authorize endpoint as a POST. The original params used in the GET call are retained as query params (as per the UI behaviour), with the new parameters required to submit the decision passed as form params.

Constructing a POST in this fashion enables the request to pass the validation for authorize requests applied by the SAPI-G filter: FapiAuthorizeRequestValidationFilter