Add support for tls_client_auth client auth method
Update config.json to configure the OB ApiClient to tls_client_auth.

Test Trusted Directory ApiClients will continue to use private_key_jwt.

The ApiClient config allows the "preferred" auth method to be requested, if it is not supported by the AS then DCR will fallback to the first supported method listed in the token_endpoint_auth_methods_supported value of the AS metadata document.